telegram_bot/requirements.txt
aiogram==3.14.0
supabase>=2.0.0
python-dotenv==1.0.0
httpx>=0.25.0
